Break up the MonaLisa related configuration in ConfigUtils
The interaction with MonaLisa configuration is now split into a configuration
source and configuration target, because MonaLisa can be both in this case.
Firstly, if the ML is present, then use it to update the application
configuration. If ML is not present, then treat is a configuration target and at
the end of the configuration initialization procedure record needed
configurations.